[Second] powf and log2f are apple only implementation

+++ second_stage/operators/selection.c
@@ -450,7 +450,7 @@ void selection_process_output (void * select_ptr, batch_p outputs) {
     /* Update pointers to use only the output array (tuples) and exclude flags and partitions */
     outputs->start += select->output_entries[2];
 
-    count = (int) powf(log2f(count), 2);
+    count = (int) pow(log2(count), 2);
 
     if (count >= 256) {
         outputs->size = count / 256 * 256;
